requirements.txt
plotly==5.18.0
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
tabulate==0.9.0
pytz==2023.3
streamlit==1.28.2 
openpyxl==3.1.2
matplotlib==3.8.0
python-dateutil==2.8.2
jira==3.5.2
tzdata==2023.3
